feat: `binance` HD derivation path

+++ packages/xchain-binance/__tests__/client.test.ts
@@ -49,10 +49,16 @@ const mockTxSend = (url: string) => {
 describe('BinanceClient Test', () => {
   let bnbClient: BinanceClient
 
+  // HDWallet: https://github.com/binance-chain/go-sdk/blob/8f0e838a5402c99cc08057a04eaece6dfd99181f/keys/hdpath.go#L20
+  // https://github.com/ebellocchia/bip_utils#complete-code-example
+
   // Note: This phrase is created by https://iancoleman.io/bip39/ and will never been used in a real-world
   const phrase = 'rural bright ball negative already grass good grant nation screen model pizza'
-  const mainnetaddress = 'bnb1zd87q9dywg3nu7z38mxdcxpw8hssrfp9e738vr'
-  const testnetaddress = 'tbnb1zd87q9dywg3nu7z38mxdcxpw8hssrfp9htcrvj'
+  const mainnetaddress_path0 = 'bnb1zd87q9dywg3nu7z38mxdcxpw8hssrfp9e738vr'
+  const mainnetaddress_path1 = 'bnb1vjlcrl5d9t8sexzajsr57taqmxf6jpmgng3gmn'
+
+  const testnetaddress_path0 = 'tbnb1zd87q9dywg3nu7z38mxdcxpw8hssrfp9htcrvj'
+  const testnetaddress_path1 = 'tbnb1zd87q9dywg3nu7z38mxdcxpw8hssrfp9htcrvj'
 
   const singleTxFee = baseAmount(37500)
   const transferFee = { type: 'base', average: singleTxFee, fast: singleTxFee, fastest: singleTxFee }
@@ -62,7 +68,7 @@ describe('BinanceClient Test', () => {
   const transferAmount = baseAmount(1000000)
 
   const phraseForTX = 'wheel leg dune emerge sudden badge rough shine convince poet doll kiwi sleep labor hello'
-  const testnetaddressForTx = 'tbnb1t95kjgmjc045l2a728z02textadd98yt339jk7'
+  const testnetaddress_path0ForTx = 'tbnb1t95kjgmjc045l2a728z02textadd98yt339jk7'
 
   const mainnetClientURL = 'https://dex.binance.org'
   const testnetClientURL = 'https://testnet-dex.binance.org'
@@ -72,17 +78,31 @@ describe('BinanceClient Test', () => {
   })
 
   afterEach(async () => {
-    bnbClient.purgeClient()
+    if (bnbClient) bnbClient.purgeClient()
   })
 
   it('should start with empty wallet', async () => {
     const bnbClientEmptyMain = new BinanceClient({ phrase, network: 'mainnet' })
     const addressMain = bnbClientEmptyMain.getAddress()
-    expect(addressMain).toEqual(mainnetaddress)
+    expect(addressMain).toEqual(mainnetaddress_path0)
 
     const bnbClientEmptyTest = new BinanceClient({ phrase, network: 'testnet' })
     const addressTest = bnbClientEmptyTest.getAddress()
-    expect(addressTest).toEqual(testnetaddress)
+    expect(addressTest).toEqual(testnetaddress_path0)
+  })
+
+  it('should support derivation index as string', async () => {
+    const bnbClientEmptyMain = new BinanceClient({ phrase, network: 'mainnet', derivationPath: '0' })
+    const addressMain = bnbClientEmptyMain.getAddress()
+    expect(addressMain).toEqual(mainnetaddress_path0)
+    const bnbClientEmptyMain_path1 = new BinanceClient({ phrase, network: 'mainnet', derivationPath: '1' })
+    expect(bnbClientEmptyMain_path1.getAddress()).toEqual(mainnetaddress_path1)
+
+    const bnbClientEmptyTest = new BinanceClient({ phrase, network: 'testnet', derivationPath: '0' })
+    const addressTest = bnbClientEmptyTest.getAddress()
+    expect(addressTest).toEqual(testnetaddress_path0)
+    const bnbClientEmptyTest_path1 = new BinanceClient({ phrase, network: 'mainnet', derivationPath: '1' })
+    expect(bnbClientEmptyTest_path1.getAddress()).toEqual(testnetaddress_path1)
   })
 
   it('throws an error passing an invalid phrase', async () => {
@@ -92,28 +112,28 @@ describe('BinanceClient Test', () => {
   })
 
   it('should have right address', async () => {
-    expect(bnbClient.getAddress()).toEqual(mainnetaddress)
+    expect(bnbClient.getAddress()).toEqual(mainnetaddress_path0)
   })
 
   it('should update net', () => {
     const client = new BinanceClient({ phrase, network: 'mainnet' })
     client.setNetwork('testnet')
     expect(client.getNetwork()).toEqual('testnet')
-    expect(client.getAddress()).toEqual(testnetaddress)
+    expect(client.getAddress()).toEqual(testnetaddress_path0)
   })
 
   it('setPhrase should return addres', () => {
-    expect(bnbClient.setPhrase(phrase)).toEqual(mainnetaddress)
+    expect(bnbClient.setPhrase(phrase)).toEqual(mainnetaddress_path0)
 
     bnbClient.setNetwork('testnet')
-    expect(bnbClient.setPhrase(phrase)).toEqual(testnetaddress)
+    expect(bnbClient.setPhrase(phrase)).toEqual(testnetaddress_path0)
   })
 
   it('should validate address', () => {
-    expect(bnbClient.validateAddress(mainnetaddress)).toBeTruthy()
+    expect(bnbClient.validateAddress(mainnetaddress_path0)).toBeTruthy()
 
     bnbClient.setNetwork('testnet')
-    expect(bnbClient.validateAddress(testnetaddress)).toBeTruthy()
+    expect(bnbClient.validateAddress(testnetaddress_path0)).toBeTruthy()
   })
 
   it('has no balances', async () => {
@@ -264,14 +284,14 @@ describe('BinanceClient Test', () => {
 
   it('should broadcast a transfer', async () => {
     const client = new BinanceClient({ phrase: phraseForTX, network: 'testnet' })
-    expect(client.getAddress()).toEqual(testnetaddressForTx)
+    expect(client.getAddress()).toEqual(testnetaddress_path0ForTx)
 
     mockGetAccount(
       testnetClientURL,
-      testnetaddressForTx,
+      testnetaddress_path0ForTx,
       {
         account_number: 0,
-        address: testnetaddressForTx,
+        address: testnetaddress_path0ForTx,
         balances: [
           {
             free: '1.00037500',
@@ -292,12 +312,16 @@ describe('BinanceClient Test', () => {
     mockNodeInfo(testnetClientURL)
     mockTxSend(testnetClientURL)
 
-    const txHash = await client.transfer({ asset: AssetBNB, recipient: testnetaddressForTx, amount: transferAmount })
+    const txHash = await client.transfer({
+      asset: AssetBNB,
+      recipient: testnetaddress_path0ForTx,
+      amount: transferAmount,
+    })
     expect(txHash).toEqual(expect.any(String))
 
-    mockGetAccount(testnetClientURL, testnetaddressForTx, {
+    mockGetAccount(testnetClientURL, testnetaddress_path0ForTx, {
       account_number: 0,
-      address: testnetaddressForTx,
+      address: testnetaddress_path0ForTx,
       balances: [
         {
           free: '1.00000000',
@@ -322,14 +346,14 @@ describe('BinanceClient Test', () => {
 
   it('should broadcast a multi transfer', async () => {
     const client = new BinanceClient({ phrase: phraseForTX, network: 'testnet' })
-    expect(client.getAddress()).toEqual(testnetaddressForTx)
+    expect(client.getAddress()).toEqual(testnetaddress_path0ForTx)
 
     mockGetAccount(
       testnetClientURL,
-      testnetaddressForTx,
+      testnetaddress_path0ForTx,
       {
         account_number: 0,
-        address: testnetaddressForTx,
+        address: testnetaddress_path0ForTx,
         balances: [
           {
             free: '1.00090000',
@@ -349,7 +373,7 @@ describe('BinanceClient Test', () => {
 
     const transactions = [
       {
-        to: testnetaddressForTx,
+        to: testnetaddress_path0ForTx,
         coins: [
           {
             asset: AssetBNB,
@@ -358,7 +382,7 @@ describe('BinanceClient Test', () => {
         ],
       },
       {
-        to: testnetaddressForTx,
+        to: testnetaddress_path0ForTx,
         coins: [
           {
             asset: AssetBNB,
@@ -367,7 +391,7 @@ describe('BinanceClient Test', () => {
         ],
       },
       {
-        to: testnetaddressForTx,
+        to: testnetaddress_path0ForTx,
         coins: [
           {
             asset: AssetBNB,
@@ -383,9 +407,9 @@ describe('BinanceClient Test', () => {
     const txHash = await client.multiSend({ transactions })
     expect(txHash).toEqual(expect.any(String))
 
-    mockGetAccount(testnetClientURL, testnetaddressForTx, {
+    mockGetAccount(testnetClientURL, testnetaddress_path0ForTx, {
       account_number: 0,
-      address: testnetaddressForTx,
+      address: testnetaddress_path0ForTx,
       balances: [
         {
           free: '1.00000000',
@@ -434,7 +458,7 @@ describe('BinanceClient Test', () => {
           blockHeight: 85905063,
           txType: 'TRANSFER',
           timeStamp: '2020-11-06T12:38:42.889Z',
-          fromAddr: testnetaddressForTx,
+          fromAddr: testnetaddress_path0ForTx,
           toAddr: 'bnb14qsnqxrjg68k5w6duq4fseap6fkg9m8fspz8f2',
           value: '1000',
           txAsset: 'BNB',
@@ -452,7 +476,7 @@ describe('BinanceClient Test', () => {
       ],
     })
 
-    const txArray = await bnbClient.getTransactions({ address: testnetaddressForTx })
+    const txArray = await bnbClient.getTransactions({ address: testnetaddress_path0ForTx })
     expect(txArray.total).toBeTruthy()
     expect(txArray.txs.length).toBeTruthy()
   })
